% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_Functions.R
\name{MasterTest}
\alias{MasterTest}
\title{Master Normality Testing Function}
\usage{
MasterTest(c, data, alpha = 0.05, j = 1)
}
\arguments{
\item{c}{Input specifying the test to run (scalar)}

\item{data}{The data of a univariate distribution for which the test statistic is computed (vector)}

\item{alpha}{The two-sided decision threshold used for hypothesis-testing (scalar)}

\item{j}{The # hypotheses tested; used to compute a Bonferonni correction, if applicable;
should remain at its default if multiple testing is not an issue (scalar)}
}
\value{
An results object specific to the test designated with the 'c' argument (list)
}
\description{
This is a master function to call the appropriate test(s) to be used in the 'Rita' function.
}
\examples{
values <- rnorm(100)
x <- MasterTest(c = 1, data = values)
}
